Lab3_InstanceSegmentation/requirements.txt
numpy
pillow
tifffile
pycocotools
